third_party/WebKit/LayoutTests/fast/dom/shadow/content-pseudo-element-overridden-2.html

<!doctype html>
<html>
<head>
<script src="resources/shadow-dom.js"></script>
<script src="../../../resources/js-test.js"></script>
</head>
<body class="bodyClass">
  <div id='sandbox'></div>
  <pre id='console'></pre>
</body>
<script>
description('Test for crbug.com/274059. Compare rules from a style in a shadow tree with ::content in a different shadow tree.');

var sandbox = document.getElementById('sandbox');
sandbox.appendChild(
    createDOM('div', {'id': 'host'},
        createShadowRoot(
            createDOM('style', {},
                document.createTextNode('::content > * { color: green; }')),
            createDOM('content', {})),
        createDOM('div', {'id': 'hostChild', 'class': 'contentClass'},
            createShadowRoot(
                createDOM('style', {},
                    document.createTextNode(':host { color: red; }')),
                createDOM('span', {},
                    document.createTextNode('Hello'))))));
shouldBe('window.getComputedStyle(getNodeInComposedTree("hostChild")).color', '"rgb(0, 128, 0)"');
sandbox.innerHTML = '';

</script>
</html>
